test: PlayerTest 내용을 NameTest로 올김
