+++ src/test/java/domain/PlayerTest.java
@@ -9,28 +9,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 public class PlayerTest {
-    @ParameterizedTest
-    @DisplayName("참여자 이름이 5자 이하고 공백이 아닌 경우 예외가 발생하지 않는다.")
-    @ValueSource(strings = {"무", "12345", "루 쿠  "})
-    void create_player(String name) {
-        assertDoesNotThrow(() -> new Player(name));
-    }
 
-    @Test
-    @DisplayName("참여자 이름이 5자를 초과하면 예외가 발생한다.")
-    void name_6() {
-        assertThatThrownBy(() -> new Player("123456"))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("이름 길이는 5자를 넘길 수 없습니다.");
-    }
-
-    @DisplayName("참여자 이름이 공백이면 에외가 발생한다.")
-    @ParameterizedTest
-    @ValueSource(strings = {"", " ", "   "})
-    void name_blank(String blankName) {
-        assertThatThrownBy(() -> new Player(blankName))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("빈 문자열 입니다.");
-    }
 
 }
